requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
docopt==0.6.2
google-api-core==1.22.4
google-api-python-client==1.6.4
google-auth==1.22.1
google-cloud-speech==2.0.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
httplib2==0.18.1
idna==2.10
libcst==0.3.13
mypy-extensions==0.4.3
numpy==1.19.2
oauth2client==4.1.3
opencv-python==4.4.0.44
pipreqs==0.4.10
playsound==1.2.2
proto-plus==1.10.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
SpeechRecognition==3.8.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.10
wincertstore==0.2
yarg==0.1.9